# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Re0Auth business API
  version: 1.0.0
  summary: The /v1 plane — what a downstream tool, and the Re0Auth frontend, actually call.
  description: |
    Re0Auth exposes **two strictly separated planes**, and this document describes
    only one of them.

    * The **protocol plane** under `/oauth/` (+ `/.well-known/`) speaks OAuth 2.0:
      form-encoded requests, `{error, error_description}` responses. Its contract
      is fixed by RFC 6749 / 7009 / 7662 / 8414 / 8628 / 9728, so it is discovered
      at `/.well-known/oauth-authorization-server` rather than restated here.
      Restating an RFC in a second place is how the two copies drift apart.
    * The **business plane** under `/v1/` is this document: JSON in, JSON out, and
      every failure rendered as RFC 9457 `application/problem+json` with a stable
      `code`.

    Two rules follow from that split and hold for every operation below:

    1. A business-plane error is **never** an OAuth error object, and vice versa.
    2. No credential is ever returned. Re0Auth holds a *binding* — an opaque,
       revocable token issued by a data source — and never hands it downstream.
       Upstream credentials (a TapTap `stoken`, say) live in the data source's own
       vault and never reach Re0Auth at all.

    ## Who calls what

    | Caller | Credential | Operations |
    |---|---|---|
    | A downstream tool (CLI, desktop, web) | Bearer access token | `federation`, `account` |
...

...

...

...

...

...
  contact:
...
servers:
...
security: []
paths:
...
  /v1/grants/{client_id}:
    delete:
      operationId: revokeGrant
      tags: [account]
      summary: Revoke an application's access to this account.
      description: |
        Deletes every token the client holds for this account. Its next call
        fails with `invalid_token` and it must ask for consent again.

        This is the **local** revocation: nothing at any data source changes, no
        binding is removed, and the account's other clients and devices keep
        working. Revoking an *upstream* credential is a different and much louder
        action, and only the data source holding it can perform it — Re0Auth does
        not have the credential to revoke.

        Idempotent. A 204 means "this client holds nothing now", which is the
        desired state either way, so a repeated request after a dropped response
        is not an error.
      parameters:
        - name: client_id
          in: path
          required: true
          description: The client whose tokens should be deleted.
          schema: { type: string }
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204':
          description: This client holds no tokens for this account. No body.
        '400':
          description: The client id was empty.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403':
          description: Missing or incorrect `X-CSRF-Token`.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /v1/sessions/sign_out:
    post:
      operationId: signOut
      tags: [session]
      summary: Destroy the browser session.
      description: |
        Destroys the session server-side; the cookie is worthless afterwards.
        This is a **local** sign-out: it revokes no data-source binding and
        touches no upstream account. The player's other devices stay signed in.
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204':
          description: The session was destroyed. No body.
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403':
          description: Missing or incorrect `X-CSRF-Token`.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /v1/account:
    delete:
      operationId: deleteAccount
      tags: [account]
      summary: Erase this account and everything it holds.
      description: |
        Irreversible. This erases the signed-in account across every store:
        data-source bindings (revoking upstream where the source can be told),
        the credentials in the vault, every issued access and refresh token, the
        browser sessions, pending requests, and finally the account row itself.

        It is **session-scoped**, so it can only erase the caller's own account,
        and it requires `acknowledge` naming the consequence, like
        `cascade_revocation`. Deletion cannot be undone, so it must not be
        something a caller can do without writing down that it means that.

        The response body reports what was removed, per store. This is not
        decoration: the account and its sessions are gone by the time the caller
        reads it, so the body is the only account of the erasure it will ever get.

        `session_scoped` in the result is false on a deployment whose session store
        cannot enumerate one account's sessions (the in-memory one). It is a
        boolean rather than a zero because "cleared no sessions" and "cannot clear
        sessions" must not look the same.
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountDeletion' }
      responses:
        '200':
          description: The account was erased. The body reports what was removed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountErasure' }
        '400':
          description: |
            A missing or wrong `acknowledge`, or a malformed body.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403':
          description: Missing or incorrect `X-CSRF-Token`.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500':
          description: |
            A store failed part-way. Every step is idempotent, so the request is
            safe to retry.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        An opaque access token. It is not a JWT and cannot be introspected by the
        client — verify it with `POST /oauth/introspect` (RFC 7662).

        DPoP is deliberately not supported in v1, and the authorization server
        metadata does not advertise it: advertising a proof-of-possession scheme
        while issuing plain bearer tokens would be a silent downgrade, which is
        worse than not offering it.
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-r0semi_session
      description: |
        The browser session, `HttpOnly` and `SameSite=Lax`. Named
        `__Host-r0semi_session` on an https deployment and `r0semi_session` on
        plain-HTTP local development. Those operations also require
        `X-CSRF-Token`, which is why this scheme alone is never sufficient for a
        write.
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: |
        The token from `GET /v1/sessions/current`. It is bound to the session, so
        a cross-site request cannot produce one. Compared in constant time.
      schema: { type: string }
...
  responses:
    Unauthenticated:
      description: |
        No usable credential. `unauthenticated` when none was presented, or
        `invalid_token` when the bearer token is unknown, revoked or expired.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: |
        The caller exceeded its per-address budget. The limiter sits in front of
        everything else, so this can appear on any operation.
      headers:
        Retry-After:
          description: Whole seconds to wait before retrying.
          schema: { type: integer, minimum: 1 }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InternalError:
      description: |
        An unexpected server-side failure. Quote `request_id` in a bug report: it
        is echoed on every response, including ones no handler produced.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      title: Problem
      description: |
        RFC 9457 Problem Details. Business plane only — the protocol plane answers
        with `{error, error_description}` instead, and the two are never mixed.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri-reference
          description: |
            A documentation URL derived from `code`, of the form
            `https://r0semi.dev/errors/<code>`.
        title:
          type: string
          description: A short human-readable summary. Stable per `code`; do not parse.
        status:
          type: integer
          description: The HTTP status, repeated for convenience.
        detail:
          type: string
          description: A human-readable explanation of this occurrence.
        instance:
          type: string
          description: The request path.
        code:
          type: string
          description: |
            The stable, machine-readable identifier for this error. Branch on
            this, never on `title` or `status`.
          enum:
            - cascade_unsupported
            - explicit_consent_required
            - internal_error
            - invalid_request
            - invalid_token
            - last_identity
            - not_acceptable
            - not_found
            - rate_limited
            - scope_not_granted
            - source_not_bound
            - source_retired
            - source_unavailable
            - unauthenticated
            - upstream_unavailable
        request_id:
          type: string
          description: |
            Matches the `X-Request-Id` response header. Present whenever the
            request reached the server.
        required_scope:
          type: string
          description: The scope that would have been needed. Set with `scope_not_granted`.
        game:
          type: string
          description: Set with `source_not_bound`.
        source:
          type: string
          description: Set with `source_not_bound`.
        bind_url:
          type: string
          format: uri
          description: |
            Where to send the player to bind the source. Set with
            `source_not_bound`.
...
    AccountDeletion:
      type: object
      required: [acknowledge]
      properties:
        acknowledge:
          type: string
          description: |
            Must be exactly `deletes_my_account`. It names the consequence, so an
            irreversible erasure cannot be requested without stating it.
          enum: [deletes_my_account]
    AccountErasure:
      type: object
      required: [result]
      properties:
        result:
          $ref: '#/components/schemas/AccountErasureResult'
    AccountErasureResult:
      type: object
      description: |
        What the erasure removed, per store. Counts are reported rather than
        asserted, because the caller cannot enumerate any of this afterwards.
      properties:
        bindings:
          type: object
          properties:
            total: { type: integer }
            revoked: { type: integer }
            failed: { type: integer }
        vault: { type: integer, description: Credentials shredded in the vault. }
        tokens: { type: integer, description: Issued access and refresh tokens deleted. }
        sessions: { type: integer }
        flows: { type: integer, description: Pending data-source bind flows deleted. }
        oidc: { type: integer, description: Pending consent requests and device authorizations deleted. }
        session_scoped:
          type: boolean
          description: |
            False when this deployment's session store cannot enumerate one
            account's sessions, so the sessions were not cleared. Distinct from
            `sessions: 0`, which means they were cleared and there were none.
...